+++ finality-provider/service/app.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	"fmt"
 	"strings"
 	"sync"
@@ -349,9 +350,35 @@ func (app *FinalityProviderApp) CreateFinalityProvider(
 		return nil, fmt.Errorf("failed to create proof-of-possession of the finality-provider: %w", err)
 	}
 
-	// TODO: query consumer chain to check if the fp is already registered
+	// Query the consumer chain to check if the fp is already registered
 	// if true, update db with the fp info from the consumer chain
 	// otherwise, proceed registration
+	resp, err := app.cc.QueryFinalityProvider(eotsPk.MustToBTCPK())
+	if err != nil {
+		if !strings.Contains(err.Error(), "the finality provider is not found") {
+			return nil, fmt.Errorf("err getting finality provider: %w", err)
+		}
+	}
+	if resp != nil {
+		app.logger.Info("finality-provider already registered on the consumer chain",
+			zap.String("eots_pk", resp.FinalityProvider.BtcPk.MarshalHex()),
+			zap.String("addr", resp.FinalityProvider.Addr),
+		)
+
+		if err := app.putFpFromResponse(resp.FinalityProvider, chainID); err != nil {
+			return nil, err
+		}
+
+		// get updated fp from db
+		storedFp, err := app.fps.GetFinalityProvider(eotsPk.MustToBTCPK())
+		if err != nil {
+			return nil, err
+		}
+
+		return &CreateFinalityProviderResult{
+			FpInfo: storedFp.ToFinalityProviderInfo(),
+		}, nil
+	}
 
 	// 3. register the finality provider on the consumer chain
 	request := &CreateFinalityProviderRequest{
@@ -557,4 +584,67 @@ func (app *FinalityProviderApp) getFpPrivKey(fpPk []byte) (*btcec.PrivateKey, er
 	}
 
 	return record.PrivKey, nil
+}
+
+// putFpFromResponse creates or updates finality-provider in the local store
+func (app *FinalityProviderApp) putFpFromResponse(fp *bstypes.FinalityProviderResponse, chainID string) error {
+	btcPk := fp.BtcPk.MustToBTCPK()
+	_, err := app.fps.GetFinalityProvider(btcPk)
+	if err != nil {
+		if errors.Is(err, store.ErrFinalityProviderNotFound) {
+			addr, err := sdk.AccAddressFromBech32(fp.Addr)
+			if err != nil {
+				return fmt.Errorf("err converting fp addr: %w", err)
+			}
+			if err := app.fps.CreateFinalityProvider(addr, btcPk, fp.Description, fp.Commission, chainID); err != nil {
+				return fmt.Errorf("failed to save finality-provider: %w", err)
+			}
+
+			app.logger.Info("finality-provider successfully saved the local db",
+				zap.String("eots_pk", fp.BtcPk.MarshalHex()),
+				zap.String("addr", fp.Addr),
+			)
+
+			return nil
+		}
+
+		return err
+	}
+
+	if err := app.fps.SetFpDescription(btcPk, fp.Description, fp.Commission); err != nil {
+		return err
+	}
+
+	if err := app.fps.SetFpLastVotedHeight(btcPk, uint64(fp.HighestVotedHeight)); err != nil {
+		return err
+	}
+
+	power, err := app.cc.QueryFinalityProviderVotingPower(btcPk, fp.Height)
+	if err != nil {
+		return fmt.Errorf("failed to query voting power for finality provider %s: %w",
+			fp.BtcPk.MarshalHex(), err)
+	}
+
+	var status proto.FinalityProviderStatus
+	switch {
+	case power > 0:
+		status = proto.FinalityProviderStatus_ACTIVE
+	case fp.SlashedBtcHeight > 0:
+		status = proto.FinalityProviderStatus_SLASHED
+	case fp.Jailed:
+		status = proto.FinalityProviderStatus_JAILED
+	default:
+		status = proto.FinalityProviderStatus_INACTIVE
+	}
+
+	if err := app.fps.SetFpStatus(btcPk, status); err != nil {
+		return fmt.Errorf("failed to update status for finality provider %s: %w", fp.BtcPk.MarshalHex(), err)
+	}
+
+	app.logger.Info("finality-provider successfully updated the local db",
+		zap.String("eots_pk", fp.BtcPk.MarshalHex()),
+		zap.String("addr", fp.Addr),
+	)
+
+	return nil
 }